Make sep_block_width a float instead of an integer, thus making it undependent from bar_height and font size.

+++ i3bar/include/common.h
@@ -47,7 +47,7 @@ struct status_block {
     bool no_separator;
 
     /* The amount of pixels necessary to render a separater after the block. */
-    uint32_t sep_block_width;
+    float sep_block_width;
 
     /* The amount of pixels necessary to render this block. These variables are
      * only temporarily used in refresh_statusline(). */
